# HOCS-2310 Running custom reports takes too long