* flush csv headers before initialising adapters
* perform convertData in parallel

I have generated CSVs from QA data using the original code and these changes. Diff suggests the results are identical. Using Chrome to time the responses gives:

-- before
695.49ms
713.03ms
680.36ms

-- after
158.51ms
114.82ms
303.93ms

Because the bulk of the data is returned in the forEach, it will be important to see if the overall time from request to download completion is still faster in one of the environments.